Refactor qr.disp_mon -> capture.disp_mon
